+++ exportplan/tests/test_helpers.py
@@ -1,7 +1,9 @@
+from unittest.mock import MagicMock, PropertyMock
+
 import pytest
 
 from exportplan import helpers
-from exportplan.tests import factories
+from exportplan.tests import factories, snapshots
 
 
 def test_country_code_iso3_to_iso2():
@@ -91,3 +93,28 @@ def test_get_unique_exportplan_name_empty():
     assert helpers.get_unique_exportplan_name(ep_no_product) == 'Export plan'
     assert helpers.get_unique_exportplan_name(ep_no_country) == 'Export plan'
     assert helpers.get_unique_exportplan_name(ep_nothing) == 'Export plan'
+
+
+def test_dictfetchall():
+    rows = [('mock_row_content_1', 1), ('mock_row_content_2', 2), ('mock_row_content_3', 3)]
+    cursor_description = (('some_str', 0), ('some_num', 1005))
+
+    mock_cursor = MagicMock()
+    mock_description = PropertyMock(return_value=cursor_description)
+    mock_cursor.fetchall.return_value = rows
+    type(mock_cursor).description = mock_description
+
+    expected = [
+        {'some_num': 1, 'some_str': 'mock_row_content_1'},
+        {'some_num': 2, 'some_str': 'mock_row_content_2'},
+        {'some_num': 3, 'some_str': 'mock_row_content_3'},
+    ]
+
+    assert helpers.dictfetchall(mock_cursor) == expected
+
+
+def test_build_query():
+    test_id = '123'
+    test_ts = '2022-07-01T15:19:11.031368'
+    query = helpers.build_query(test_id, test_ts)
+    assert query.strip() == snapshots.BUILD_QUERY_SNAPSHOT.strip()